Adds isValid method to AudioFile, adds static parse method to Range, and renames transitionDelay to delayRange
The audio file name validity check is now made by the public static isValid(String) method in AudioFile to allow
checking for validity without creating an instance.
The constructor for the simple Range class was made public and a public static parse(String) method was added
to parse the result of the toString method which was also added.
The transitionDelay property was renamed to delayRange to improve readability and understanding of the type
for when it is serialized to JSON and interacted with through javascript.