Adds back AudioFile path functionality and adds equals and hashCode implementations
The ability to create AudioFile instances for files including the file path was added back
since this is needed for messages sent from the server to web browser client.
The AudioFile and Range classes were updated to include equals and hashCode
implementations so that they can be stored and compared when creating AudioMessages.